refactor(canvas): 重构画布编辑器代码结构提升可维护性
- 优化了 InfiniteCanvas 组件结构，整理导入与变量声明
- 维护拖拽文件类型、菜单状态及视图状态的逻辑
- 简化节点添加、拖拽及连接处理逻辑，保持代码风格一致
- 保持文件上传和资源同步流程，增强错误处理
- 调整快捷键和自动保存逻辑，提升用户体验
- 保持UI元素及交互提示的一致性与响应性
