+++ frontend/src/app/theater/[id]/page.tsx
@@ -854,7 +854,7 @@ function InfiniteCanvas() {
           </Panel>
           
           <Panel position="top-left" className="m-4 z-50">
-            <div className="flex items-center bg-card border border-border/50 shadow-sm rounded-lg p-1 gap-1 pointer-events-auto">
+            <div className="flex items-center bg-card border border-border/50 rounded-lg p-1 gap-1 pointer-events-auto">
               <Button variant="ghost" size="icon" className="h-8 w-8 text-muted-foreground hover:text-foreground" onClick={() => router.push('/')} title={t('canvas.back')}>
                 <ArrowLeft className="w-4 h-4" />
               </Button>
